--- astarte.py
@@ -1,13 +1,0 @@
-from src.person.person import Person
-
-
-class Astarte(Person):
-    def __init__(self, name: str, planet: 'Planet' or dict, founding: int) -> None:
-        planet = Planet(planet['planet_name'],planet['planet_type'])
-        super().__init__(name, planet)
-        self.__founding = founding
-
-
-    @property
-    def founding(self) -> int:
-        return self.__founding
