+++ src/enumeration/enumeration.py
@@ -1,0 +1,22 @@
+from enum import Enum
+
+
+class Status(Enum):
+    ALIVE = 'Alive'
+    DEAD = 'Dead'
+    UNKNOWN = 'Unknown'
+
+
+
+class PlanetType(Enum):
+    AGRI = 'Agri'
+    CIVILISED = 'Civilised'
+    DAEMON = 'Daemon'
+    DEAD = 'Dead'
+    DEATH = 'Death'
+    FERAL = 'Feral'
+    FEUDAL = 'Feudal'
+    FORGE = 'Forge'
+    FRONTIER = 'Frontier'
+    HIVE = 'Hive'
+
